♻️ operator: dedup org namespace derivation + extract boundary types
Cleanup pass on the ClusterTenant provisioning code (PR #50 follow-up):

- Namespace derivation lived in three places (shared-cluster provisioner's
  hardcoded opencrane- prefix, the org-domain provisioner re-deriving
  ${namespacePrefix}${org}, and a CLUSTER_TENANT_NAMESPACE_PREFIX config that
  could silently diverge from the hardcoded boundary contract). Collapse to ONE
  source: the reconciler derives boundNamespace via _NamespaceForOrg and passes
  it through OrgDomainProvisionRequest. Removed the namespacePrefix from the
  provisioner config and the now-unused clusterTenantNamespacePrefix config
  field + env (no compat shim).
- Moved BoundaryProvisionResult + ClusterTenantReconcilePhase out of
  shared-cluster.provisioner.ts into a sibling .types.ts, matching the
  org-domain-provisioner.types.ts convention; the logic file now carries only
  behaviour. No re-export from the old location.
- plan.md: recorded three follow-ups — DOMAIN.T1 (k8s-native DNS via
  external-dns instead of the direct @google-cloud/dns binding), DOMAIN.T2
  (wire deprovisionOrgDomain on CR delete; external Cloud DNS records currently
  leak), DOMAIN.T3 (collapse the dead control-plane provisioner runtime paths).

operator: tsc clean, 134 tests pass.
