+++ apps/operator/src/__tests__/runtime-planes/drift-repairer.test.ts
@@ -19,7 +19,6 @@ function _buildConfig(overrides: Partial<OpenClawTenantOperatorConfig> = {}): Op
     dnsManagedZone: "",
     certManagerIssuerName: "opencrane-issuer",
     certManagerIssuerKind: "ClusterIssuer",
-    clusterTenantNamespacePrefix: "opencrane-",
     ingressTlsEnabled: false,
     ingressTlsSecretName: "opencrane-wildcard-tls",
     gatewayPort: 18789,
